[REFACTOR] chatlist page에서 api 통신로직 서버컴포넌트로 이동
