+++ src/lib/api/SupabaseApi.ts
@@ -5,6 +5,7 @@ import type { FlirtingListInNotificationType, FlirtingListType } from '@/types/f
 import type { MessageType, SendMessageType, SpecificSubscribeFlirtingListCallbackType } from '@/types/realTimeType';
 import type { ChatListType } from '@/types/realTimeType';
 import { format } from 'date-fns';
+import { LastMessageArrayType } from '@/types/lastMessageArrayType';
 
 // const client = createClient(process.env.NEXT_PUBLIC_SUPABASE_URL || '', process.env.NEXT_PUBLIC_SERVICE_KEY || '');
 
@@ -284,6 +285,28 @@ export async function subscribeChatList(callback: SpecificSubscribeFlirtingListC
       callback
     )
     .subscribe();
+}
+
+// chat list에서 빼옴
+export async function fetchLastMessages(roomIds: string[]): Promise<LastMessageArrayType> {
+  let lastMessageArray = [];
+
+  for (let i = 0; i < roomIds.length; i++) {
+    const { data: getLastMessage, error } = await supabase
+      .from('message')
+      .select('*')
+      .eq('subscribe_room_id', roomIds[i])
+      .order('created_at', { ascending: false })
+      .limit(1)
+      .maybeSingle();
+
+    if (error) {
+      console.error('Error fetching last message:', error);
+      throw new Error('Error fetching last message');
+    }
+    lastMessageArray.push(getLastMessage);
+  }
+  return lastMessageArray;
 }
 
 export async function untrackChatList() {
